% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_animate_path.R
\name{animate_path}
\alias{animate_path}
\title{CuH2 visualization CLI}
\usage{
animate_path()
}
\arguments{
\item{dir}{The directory path where the files are stored. \link{required}}

\item{ground}{The path to the file containing the base system configuration (ground state). \link{required}}

\item{pattern}{The file pattern to match in the directory. Default is "neb_path_\%03d.con".}

\item{rangeidx}{A comma-separated string specifying the range of file indices to include. Example: "0,10". Default is "0,10".}

\item{filename}{The filename for the output animation, without the file type. Default is "neb_cuh2".}

\item{clip_max}{The maximum energy of the contour plot. Default is 5.}

\item{clip_min}{The minimum energy of the contour plot. Default is 0.}

\item{width}{The width of the output animation in pixels. Default is 800.}

\item{height}{The height of the output animation in pixels. Default is 800.}

\item{fps}{The frames per second for the output animation. Default is 60.}

\item{res}{The resolution for the output animation. Default is 150.}

\item{duration}{The duration of the output animation in seconds. Default is 10.}

\item{hh_range}{A comma-separated string specifying the range for H-H distances. Example: "0.4,3.2". Default is "0.4,3.2".}

\item{hcu_range}{A comma-separated string specifying the range for H-Cu distances. Example: "-0.05,5.1". Default is "-0.05,5.1".}
}
\value{
Prints the success message and saves the generated animation to the specified file.
}
\description{
This function provides a command-line interface for the CuH2 animation
tools. It requires a directory path and a ground state configuration file.
Additionally, it accepts optional arguments for file pattern, range of file
indices to include, and animation parameters.
}
\details{
The "ground state configuration" refers to a configuration file that describes
any state of the system, typically a reactant or product configuration.
This configuration is used as a base for generating the energy surface that
guides the animation.
}
\examples{
animate_path(dir = "/path/to/files", ground = "/path/to/ground_file.con")
}
