% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotaids.R
\name{cuh2vizR_generate_animation}
\alias{cuh2vizR_generate_animation}
\title{Generate and plot energy contours with animation}
\usage{
cuh2vizR_generate_animation(
  dfx,
  df_list,
  clip_max = 5,
  clip_min = 0,
  filename = "neb_cuh2",
  startrange = 0,
  duration = 10,
  fps = 60,
  width = 800,
  height = 800,
  res = 150
)
}
\arguments{
\item{dfx}{A data frame containing the columns 'hh_dist', 'hcu_dist' and 'energy'.}

\item{df_list}{A list of data frames for each iteration.}

\item{clip_max}{The maximum energy level. Default is 5.}

\item{clip_min}{The minimum energy level. Default is 0.}

\item{filename}{The name of the output file for the animation.}

\item{duration}{The duration of the animation. Default is 10 seconds.}

\item{fps}{The number of frames per second in the animation. Default is 60.}

\item{width}{The width of the output gif in pixels. Default is 800.}

\item{height}{The height of the output gif in pixels. Default is 800.}

\item{res}{The resolution of the output gif in dpi. Default is 150.}
}
\description{
This function takes a data frame and plots the energy contours
using ggplot2 and gganimate. The energy levels can be clipped with the \code{clip_max}
and \code{clip_min} parameters. The animation is saved as a GIF file.
}
\examples{
\dontrun{
cuh2vizR_generate_animation(
  dfx,
  df_list,
  clip_max = 5,
  clip_min = 0,
  filename = "neb_cuh2.gif",
  duration = 10,
  fps = 60,
  width = 800,
  height = 800,
  res = 150
)
}
}
