% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotaids.R
\name{cuh2vizR_get_energy_contours}
\alias{cuh2vizR_get_energy_contours}
\title{Generate and plot energy contours}
\usage{
cuh2vizR_get_energy_contours(a_df, clip_max = 5, clip_min = 0)
}
\arguments{
\item{clip_max}{The maximum energy level. Default is 5.}

\item{clip_min}{The minimum energy level. Default is 0.}

\item{df}{A data frame containing the columns 'hh_dist', 'hcu_dist' and 'energy'.}
}
\value{
A ggplot object.
}
\description{
This function takes a data frame and plots the energy contours
using ggplot2. The energy levels can be clipped with the \code{clip_max}
and \code{clip_min} parameters.
}
\examples{
\dontrun{
cuh2vizR_get_energy_contours(df, clip_max = 5, clip_min = 0)
}
}
