% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuh2vizR-package.R
\name{cuh2vizR_get_energy_df}
\alias{cuh2vizR_get_energy_df}
\title{Generate energy dataframe}
\usage{
cuh2vizR_get_energy_df(
  a_df,
  hcu_dists = seq(-0.05, 5.1, length.out = 60),
  hh_dists = seq(0.4, 3.2, length.out = 60)
)
}
\arguments{
\item{hcu_dists}{A vector of H-Cu distances for which the energies are to be computed.}

\item{hh_dists}{A vector of H-H distances for which the energies are to be computed.}

\item{df}{A data frame containing the atom data required by 'get_energy'.}
}
\value{
A data frame with columns 'hcu_dist', 'hh_dist' and 'energy'.
}
\description{
This function generates a data frame that represents the energy
levels based on the distances specified by 'hcu_dists' and 'hh_dists'.
The energy levels are computed by the function 'get_energy'.
}
\examples{
\dontrun{
cuh2vizR_get_energy_df(df, hcu_dists = seq(-0.05, 5, length.out = 60), hh_dists = seq(0.4, 3, length.out = 60))
}
}
